requirements.txt
google-cloud-aiplatform>=1.78.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
google-adk>=1.0.0
google-genai>=1.65.0
python-dotenv>=1.0.0
pytest 
pytest-asyncio